requirements.txt
django
django-environ
djangorestframework
black